pyproject.toml
[tool.poetry]
name = "attendencemanager-email-server"
version = "0.1.0"
description = "This sends an email to registered user when they are marked absent in any period by monitoring their student portal."
authors = ["Devasheesh Mishra <devasheeshmishra4@gmail.com>"]
readme = "ReadMe.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pytesseract = "^0.3.13"
beautifulsoup4 = "4.12.3"
httpx = "^0.27.2"
pydantic = "^2.9.1"
pandas = "^2.2.2"
lxml = "^5.3.0"
sqlalchemy = "^2.0.35"
psycopg = {extras = ["binary"], version = "^3.2.3"}
sqlalchemy-utils = "^0.41.2"
alembic = "^1.13.3"
pydantic-settings = "^2.5.2"
azure-communication-email = "^1.0.0"
psycopg2 = "^2.9.9"
apscheduler = "^3.10.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
